Revert "add speech understanding (#14)" (#15)
This reverts commit e29b5d02907ef809b1df8c3279a4589618661450.
