+++ dataset.py
@@ -79,7 +79,6 @@ class SpeechDataset(Dataset):
         ids_audio = [0] * self.config.max_speech_token_size
         tgt_audio = [IGNORE_TOKEN_ID] * len(ids_audio)
         chat = [{"role": "user", "content": "Transcribe the speech"}]
-        # chat = [{"role": "user", "content": "Translate the audio to English"}]
         if self.inference:
             kwargs = {'add_generation_prompt': True}
         else:
